BatchedTransformations

Documentation for BatchedTransformations.

BatchedTransformations.AbstractLinearMaps

BatchedTransformations.AffineMaps

BatchedTransformations.ComposedTransformations

BatchedTransformations.InverseTransformations

BatchedTransformations.LinearMaps

BatchedTransformations.RigidTransformations

BatchedTransformations.Rotations

BatchedTransformations.Transformations

BatchedTransformations.Translations

BatchedTransformations.compose

BatchedTransformations.transform

BatchedTransformations.AbstractLinearMaps — Type

AbstractLinearMaps <: Transformations

BatchedTransformations.AffineMaps — Method

AffineMaps(translations::Translations, linear::AbstractLinearMaps)

BatchedTransformations.ComposedTransformations — Type

ComposedTransformations{Outer<:Transformations,Inner<:Transformations}

A ComposedTransformations contains two transformations t2 and t1 that are composed. It can be constructed with compose(t2, t1), t2 ∘ t1, and t2(t1), where t1 is the transformation to be applied first, and t2 second.

BatchedTransformations.InverseTransformations — Type

InverseTransformations{T<:Transformations}

A InverseTransformations represents a lazy inverse of a Transformations t.

inverse(t) is a lazy inverse that defaults to inv(t) when evaluated. transform(inverse(t), x) is equivalent to inverse_transform(t, x). This allows for specialized inverse transform implementations that don't require the inverse to be computed explicitly.

BatchedTransformations.LinearMaps — Type

LinearMaps{A<:AbstractArray} <: AbstractLinearMaps

BatchedTransformations.RigidTransformations — Method

RigidTransformations(translations::Translations, rotations::Rotations)

BatchedTransformations.Rotations — Type

Rotations{A<:AbstractArray} <: AbstractLinearMaps

BatchedTransformations.Transformations — Type

Transformations

A Transformations contains a batch of transformations that can be applied to an array. A Transformations t can be applied to x with transform(t, x), t ∘ x, and t(x).

BatchedTransformations.Translations — Type

Translations{A<:AbstractArray} <: Transformations

BatchedTransformations.compose — Method

compose(t2, t1)
t2 ∘ t1
t2(t1)

BatchedTransformations.transform — Method

transform(t, x)
t ∘ x
t(x)