BatchedTransformations

Documentation for BatchedTransformations.

BatchedTransformations.Composed

BatchedTransformations.Identity

BatchedTransformations.Inverse

BatchedTransformations.Transformation

BatchedTransformations.compose

BatchedTransformations.transform

BatchedTransformations.Composed — Type

Composed{Outer<:Transformation,Inner<:Transformation}

A Composed contains two transformations outer and inner that are composed, where inner gets applied first, and then outer.. It can be constructed with compose(outer, inner) or outer ∘ inner, unless the compose function is overloaded for the specific types.

BatchedTransformations.Identity — Type

Identity <: Transformation

BatchedTransformations.Inverse — Type

Inverse{T<:Transformation} <: Transformation

An Inverse represents a lazy inverse of a Transformation t.

inverse(t) is a lazy inverse that defaults to inv(t) when evaluated. transform(inverse(t), x) is equivalent to inverse_transform(t, x). This allows for specialized inverse transform implementations that don't require the inverse to be computed explicitly.

BatchedTransformations.Transformation — Type

Transformation

An abstract type whose concrete subtypes contain batches of transformations that can be applied to an array. A Transformation t can be applied to x with transform(t, x), t * x, and t(x).

BatchedTransformations.compose — Method

compose(t2, t1)
t2 ∘ t1

BatchedTransformations.transform — Method

transform(t, x)
t * x
t(x)